# Improve performance of `var` variable declarations