--- a/docs/CHANGELOG.md
+++ b/docs/CHANGELOG.md
@@ -7,6 +7,8 @@ Version 3.41.0-eisop2 (December ?, 2023)
 
 **Closed issues:**
 
+typetools#6373.
+
 
 Version 3.41.0-eisop1 (December 5, 2023)
 ----------------------------------------