Some timing:

Before this PR:
- using `var`, with init: 165s
- using `var`, assume init: 90s
- using `Object`, with init: 50s
- using `Object`, assume init: 33s

After this PR:
- using `var`, with init: 105s
- using `var`, assume init: 60s
- using `Object`, with init: 45s
- using `Object`, assume init: 30s

There is still a large difference between using `var` and using `Object`, which we should investigate more.

typetools, at https://github.com/typetools/checker-framework/commit/f58c9d93d562c923d2fea17fec978eb2b8040858 (commit before https://github.com/typetools/checker-framework/commit/dc0e62c36818cf76762a4357bbb597d559cbbb28):
- using `var`: 43s
- using `Object`: 42s

typetools, at https://github.com/typetools/checker-framework/commit/dc0e62c36818cf76762a4357bbb597d559cbbb28:
- using `var`: 90s
- using `Object`: 42s

typetools, at https://github.com/typetools/checker-framework/commit/ccec69e53393e3c6dbbbc7bc53f8ad60be6fd679 (this test case crashes on typetools HEAD):
- using `var`: 115s
- using `Object`: 43s

So, this PR gets eisop in line with typetools (which is good, given eisop splits init and nullness), but doesn't undo the performance hit from improved `var` support.

Co-authored-by: Liam Miller-Cushon <cushon@google.com>