@wmdietl 2023-12-20
@flo2702 Can you have a look at the performance difference between using the Initialization Checker and disabling it? Maybe you see something else that impacts performance?