@wmdietl 2023-12-31
The newest changes in https://github.com/eisop/checker-framework/pull/657/commits/6109041c90ead352793e497da2ad6f21465c2ad3 re-enable caching in `getAnnotatedTypeLhs` for variables declared using `var`.
This only has an effect on the initializer expression, avoiding that its type is re-computed multiple times.
The expanded tests in https://github.com/eisop/checker-framework/pull/657/commits/43babac00e8dd80ec94234195c8bae36da3b22c7 make sure that re-assignment to such a variable doesn't cause re-computation.

With initialization, using both `Object` and `var`, the time is now roughly equal, at around 45s.